# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Inventra API
  version: 0.1.0
  description: |
    Asset management API for Inventra (see docs/PRD.md, docs/DATABASE.md).

    **Auth:** Bearer JWT. Obtain tokens via `POST /api/v1/auth/login`, then send
    `Authorization: Bearer <access_token>`. Access tokens are short-lived;
    rotate with `POST /api/v1/auth/refresh`. `POST /api/v1/auth/logout` revokes
    the current access token and the supplied refresh token.
  contact:
    name: Inventra
    url: https://github.com/ragbuaj/inventra
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/requests:
    post:
      tags: [Requests]
      operationId: submitRequest
      summary: Submit an approval request
      description: Creates a new approval request and routes it to the first eligible approver. Requires the `request.create` permission.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitRequest" }
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: No matching approval threshold configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [Requests]
      operationId: listRequests
      summary: List approval requests (data-scoped)
      description: Returns requests within the caller's data scope. No special permission beyond authentication required for listing. Fields `amount`/`payload`/`reason` may be omitted per role field-permissions.
      security:
        - bearerJWT: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
          description: Filter by request status.
        - name: type
          in: query
          schema:
            type: string
            enum: [asset_create, asset_disposal, asset_transfer, assignment, maintenance, valuation_exclusion]
          description: Filter by request type.
        - name: mine
          in: query
          schema: { type: boolean, default: false }
          description: |
            When `true`, returns only the caller's own submitted requests (filtered by
            requester) and bypasses office data-scope filtering — a caller can always see
            their own requests regardless of scope configuration.
        - { name: limit, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0, minimum: 0 } }
      responses:
        "200":
          description: A page of requests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/approval-thresholds:
    get:
      tags: [Approval Thresholds]
      operationId: listThresholds
      summary: List approval thresholds
      description: Returns all configured approval thresholds. Requires the `approval.config.manage` permission.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: All thresholds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThresholdList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Approval Thresholds]
      operationId: createThreshold
      summary: Create an approval threshold
      description: Requires the `approval.config.manage` permission.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ThresholdRequest" }
      responses:
        "201":
          description: Created threshold
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Threshold" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Duplicate threshold configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/approval-thresholds/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Approval Thresholds]
      operationId: updateThreshold
      summary: Update an approval threshold
      description: |
        Replaces mutable threshold fields (`amount_from`, `amount_to`, `required_level`,
        `step_order`, `is_active`). `request_type` is immutable after creation.
        Requires the `approval.config.manage` permission.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ThresholdRequest" }
      responses:
        "200":
          description: Updated threshold
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Threshold" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Approval Thresholds]
      operationId: deleteThreshold
      summary: Delete an approval threshold
      description: Requires the `approval.config.manage` permission.
      security:
        - bearerJWT: []
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Malformed request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or revoked token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but lacking the required permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
      examples:
        - error: invalid email or password
...
    SubmitRequest:
      type: object
      description: Body for POST /requests.
      required: [type, amount, office_id]
      properties:
        type:
          type: string
          enum: [asset_create, asset_disposal, valuation_exclusion]
        amount:
          type: string
          description: >-
            Decimal string (IDR). Determines which threshold steps apply.
            For asset_create it must numerically equal payload.purchase_cost
            (or 0 when the payload has no purchase_cost); a mismatch is
            rejected with 400 so the maker cannot understate the amount to
            route the request through a lower approval band.
        office_id:
          type: string
          format: uuid
          description: Office initiating the request.
        target_id:
          type: [string, "null"]
          format: uuid
          description: Optional reference to the target entity (e.g. asset id for disposal).
        payload:
          type: [object, "null"]
          additionalProperties: true
          description: >-
            Arbitrary request payload (e.g. asset details for asset_create —
            see the AssetCreatePayload schema for its shape).
        reason:
          type: [string, "null"]
          description: Optional narrative justification.
...
    Request:
      type: object
      description: An approval request record.
      required: [id, type, status, current_step, requested_by_id, created_at]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [asset_create, asset_disposal, asset_transfer, assignment, maintenance, valuation_exclusion]
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        amount:
          type: string
          description: Decimal string (IDR).
        current_step: { type: integer }
        office_id: { type: [string, "null"], format: uuid }
        target_id: { type: [string, "null"], format: uuid }
        target_entity: { type: [string, "null"] }
        reason: { type: [string, "null"] }
        requested_by_id: { type: string, format: uuid }
        requested_by_name:
          type: [string, "null"]
          description: Resolved maker display name (null if the user was deleted).
        requested_by_role:
          type: [string, "null"]
          description: Resolved maker role name.
        office_name:
          type: [string, "null"]
          description: Resolved originating office name.
        decided_by_id: { type: [string, "null"], format: uuid }
        decision_note: { type: [string, "null"] }
        created_at: { type: [string, "null"], format: date-time }
...
    RequestList:
      type: object
      required: [data, total, limit, offset]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Request" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    Threshold:
      type: object
      description: A single approval-threshold rule mapping request type + amount range to a required approver level.
      required: [id, request_type, amount_from, required_level, step_order, is_active]
      properties:
        id: { type: string, format: uuid }
        request_type:
          type: string
          enum: [asset_create, asset_disposal, asset_transfer, assignment, maintenance, valuation_exclusion]
        amount_from:
          type: string
          description: Lower bound (IDR decimal string, inclusive).
        amount_to:
          type: [string, "null"]
          description: Upper bound (IDR decimal string, inclusive). Null means no upper bound.
        required_level:
          type: string
          enum: [office, office_subtree, wilayah, pusat]
        step_order: { type: integer }
        is_active: { type: boolean }
        created_at: { type: [string, "null"], format: date-time }
    ThresholdRequest:
      type: object
      description: Body for POST/PUT /approval-thresholds.
      required: [request_type, amount_from, required_level, step_order]
      properties:
        request_type:
          type: string
          enum: [asset_create, asset_disposal, asset_transfer, assignment, maintenance, valuation_exclusion]
          description: Immutable after creation (ignored on PUT).
        amount_from:
          type: string
          description: Lower bound (IDR decimal string).
        amount_to:
          type: [string, "null"]
          description: Upper bound (IDR decimal string). Omit or null for no upper bound.
        required_level:
          type: string
          enum: [office, office_subtree, wilayah, pusat]
        step_order:
          type: integer
          description: Order of this step within the threshold chain (1-based).
        is_active:
          type: boolean
          default: false
    ThresholdList:
      type: object
      required: [data, total]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Threshold" } }
        total: { type: integer }
...